+++ src/index.ts
@@ -78,6 +78,7 @@ export { default as Tree } from '@/package/components/tree.vue';
 export { default as TreeSelect } from '@/package/components/tree-select.vue';
 export { default as Breadcrumbs } from '@/package/components/breadcrumbs.vue';
 export { default as Divider } from '@/package/components/divider.vue';
+export { default as PageHeader } from '@/package/components/page-header.vue';
 export { default as VirtualScroller } from '@/package/components/virtual-scroller.vue';
 export type { DataTableColumn } from '@/package/components/data-table.vue';
 export type {